+++ src/unpoly/element.js
@@ -1556,6 +1556,23 @@ up.element = (function() {
     }
   }
 
+  // Ponyfill for Element.prototype.moveBefore()
+  // https://developer.mozilla.org/en-US/docs/Web/API/Element/moveBefore
+  function moveBefore(parent, movedElement, referenceElement) {
+    let fn = parent.moveBefore || parent.insertBefore
+    fn.call(parent, movedElement, referenceElement)
+  }
+
+  function preservingAppend(parent, newNode) {
+    moveBefore(parent, newNode, null)
+  }
+
+  function preservingReplace(oldElement, newElement) {
+    let parent = oldElement.parentElement
+    moveBefore(parent, newElement, oldElement)
+    oldElement.remove()
+  }
+
   return {
     subtree,
     subtreeFirst,
@@ -1624,5 +1641,7 @@ up.element = (function() {
     matchSelectorMap,
     elementLikeMatches,
     documentPosition,
+    preservingAppend,
+    preservingReplace,
   }
 })()